# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Naples — track record API",
    "description": "\nRead-only view over the Serie A prediction track record.\n\n**The production model is M1**: the bookmaker's opening line with the margin\nremoved. Fair odds and overround are exposed because they are descriptive;\nexpected value, stakes and betting advice are not exposed at all — an EV\ncomputed from M1's probabilities against the very odds they come from is\ncircular by construction, and measured to be negative on every single bet.\n\nWriting happens only from local scheduled commands, never over HTTP.\n",
    "version": "1.0.0"
  },
  "paths": {
    "/api/season/{season}": {
      "get": {
        "summary": "Season headline figures",
        "operationId": "get_season_api_season__season__get",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rounds/{season}": {
      "get": {
        "summary": "Every round with its state and archived RPS",
        "operationId": "get_rounds_api_rounds__season__get",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Round"
                  },
                  "title": "Response Get Rounds Api Rounds  Season  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rounds/{season}/{matchday}": {
      "get": {
        "summary": "Matches of one round",
        "operationId": "get_round_api_rounds__season___matchday__get",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Season"
            }
          },
          {
            "name": "matchday",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Matchday"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Match"
                  },
                  "title": "Response Get Round Api Rounds  Season   Matchday  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matches/{season}": {
      "get": {
        "summary": "All matches of a season, filterable",
        "operationId": "get_matches_api_matches__season__get",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Season"
            }
          },
          {
            "name": "team",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Home or away, exact name",
              "title": "Team"
            },
            "description": "Home or away, exact name"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "predicted | resolved | invalid",
              "title": "Status"
            },
            "description": "predicted | resolved | invalid"
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "YYYY-MM-DD, inclusive",
              "title": "From"
            },
            "description": "YYYY-MM-DD, inclusive"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "YYYY-MM-DD, inclusive",
              "title": "To"
            },
            "description": "YYYY-MM-DD, inclusive"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Match"
                  },
                  "title": "Response Get Matches Api Matches  Season  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/track-record/{season}": {
      "get": {
        "summary": "Cumulative RPS series and calibration",
        "operationId": "get_track_record_api_track_record__season__get",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/standings/{season}": {
      "get": {
        "summary": "League table from played matches",
        "description": "The table, computed from results — a fact, not a forecast.\n\nTied teams are separated head-to-head first, as Serie A does, and not by\ngoal difference: the shortcut produces a table that looks right and is\nwrong exactly when the standings matter.",
        "operationId": "get_standings_api_standings__season__get",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Standings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/picks/{season}": {
      "get": {
        "summary": "The canonical M1 selections — no threshold to move",
        "description": "The main line, and it takes no odds parameter on purpose.\n\nThese are the selections `predict_round` prints on Friday and the weekly\nreport shows, with `config.QUOTA_MINIMA_SELEZIONE` as the declared floor.\nThe band explorer (`/api/selections`) sits beside them as an adjustable\nsecond reading: if a threshold chosen in the interface could redefine the\nmain line, the track record would be measuring one thing while the\ndashboard advertised another.",
        "operationId": "get_picks_api_picks__season__get",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Season"
            }
          },
          {
            "name": "matchday",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 38,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Matchday"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Picks"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/selections/{season}": {
      "get": {
        "summary": "Markets priced under a ceiling on the odds",
        "operationId": "get_selections_api_selections__season__get",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Season"
            }
          },
          {
            "name": "max_odds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 100.0,
              "exclusiveMinimum": 1.0,
              "description": "Ceiling on the FAIR odds, 1/p",
              "default": 1.4,
              "title": "Max Odds"
            },
            "description": "Ceiling on the FAIR odds, 1/p"
          },
          {
            "name": "min_odds",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100.0,
                  "exclusiveMinimum": 1.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Floor on the fair odds. Below it the market pays too little to be worth a bet; the band is the useful filter",
              "default": 1.3,
              "title": "Min Odds"
            },
            "description": "Floor on the fair odds. Below it the market pays too little to be worth a bet; the band is the useful filter"
          },
          {
            "name": "matchday",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 38,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Matchday"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Selections"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Picks": {
        "properties": {
          "season": {
            "type": "string",
            "title": "Season"
          },
          "matchday": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matchday"
          },
          "min_odds": {
            "type": "number",
            "title": "Min Odds",
            "description": "The declared floor, config.QUOTA_MINIMA_SELEZIONE. Not a request parameter: the main line cannot be re-cut per call"
          },
          "matches_total": {
            "type": "integer",
            "title": "Matches Total",
            "default": 0
          },
          "most_probable": {
            "items": {
              "$ref": "#/components/schemas/Selection"
            },
            "type": "array",
            "title": "Most Probable",
            "default": []
          },
          "most_probable_resolved": {
            "type": "integer",
            "title": "Most Probable Resolved",
            "default": 0
          },
          "most_probable_won": {
            "type": "integer",
            "title": "Most Probable Won",
            "default": 0
          },
          "with_min_odds": {
            "items": {
              "$ref": "#/components/schemas/Selection"
            },
            "type": "array",
            "title": "With Min Odds",
            "default": []
          },
          "with_min_odds_resolved": {
            "type": "integer",
            "title": "With Min Odds Resolved",
            "default": 0
          },
          "with_min_odds_won": {
            "type": "integer",
            "title": "With Min Odds Won",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "season",
          "min_odds"
        ],
        "title": "Picks",
        "description": "The canonical selections. No threshold field to move, by design.\n\n`most_probable` is the most likely market of each match, whatever it pays;\n`with_min_odds` is the most likely among those paying at least `min_odds`,\nwhich comes from `config.QUOTA_MINIMA_SELEZIONE` and not from the request.\nBoth are the rules the project has used since the first registered round."
      },
      "Round": {
        "properties": {
          "season": {
            "type": "string",
            "title": "Season"
          },
          "matchday": {
            "type": "integer",
            "title": "Matchday"
          },
          "first_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Date"
          },
          "last_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Date"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "matches": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matches"
          },
          "matches_with_odds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matches With Odds"
          },
          "matches_predicted": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matches Predicted"
          },
          "matches_played": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matches Played"
          },
          "matches_resolved": {
            "type": "integer",
            "title": "Matches Resolved",
            "default": 0
          },
          "rps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rps",
            "description": "Archived RPS, null unless closed"
          },
          "rps_market": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rps Market"
          },
          "closed": {
            "type": "boolean",
            "title": "Closed"
          }
        },
        "type": "object",
        "required": [
          "season",
          "matchday",
          "closed"
        ],
        "title": "Round",
        "description": "A matchday and where it stands.\n\n`status` is translated from the project's single state machine\n(`rounds.stato_giornate`), never recomputed here: future, open, predicted,\npartially_predicted, played, closed."
      },
...
      "SeasonSummary": {
        "properties": {
          "season": {
            "type": "string",
            "title": "Season"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "rounds_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rounds Total"
          },
          "rounds_closed": {
            "type": "integer",
            "title": "Rounds Closed",
            "default": 0
          },
          "matches_predicted": {
            "type": "integer",
            "title": "Matches Predicted",
            "default": 0
          },
          "matches_resolved": {
            "type": "integer",
            "title": "Matches Resolved",
            "default": 0
          },
          "matches_invalid": {
            "type": "integer",
            "title": "Matches Invalid",
            "default": 0
          },
          "rps_production": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rps Production"
          },
          "rps_backtest_reference": {
            "type": "number",
            "title": "Rps Backtest Reference"
          },
          "hits": {
            "type": "integer",
            "title": "Hits",
            "default": 0
          },
          "hits_denominator": {
            "type": "integer",
            "title": "Hits Denominator",
            "default": 0
          },
          "hit_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hit Rate"
          },
          "sample_significant": {
            "type": "boolean",
            "title": "Sample Significant"
          },
          "significance": {
            "$ref": "#/components/schemas/Significance"
          }
        },
        "type": "object",
        "required": [
          "season",
          "model_version",
          "rps_backtest_reference",
          "sample_significant",
          "significance"
        ],
        "title": "SeasonSummary",
        "description": "Headline figures. Hits keep numerator and denominator apart on purpose: a\nbare rate hides how thin the sample is, and right now it is very thin.\n\n`rps_backtest_reference` is 0.1881, the market's RPS on the test set. A\npersistent gap from it would not be a model getting things wrong — the\nmodel *is* the opening line — it would be the production pipeline behaving\ndifferently from the backtest."
      },
...
      "Significance": {
        "properties": {
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method",
            "description": "cluster-bootstrap | insufficient-rounds"
          },
          "rounds": {
            "type": "integer",
            "title": "Rounds"
          },
          "matches": {
            "type": "integer",
            "title": "Matches"
          },
          "tolerance": {
            "type": "number",
            "title": "Tolerance",
            "description": "RPS gap from the backtest worth detecting (config.TRACK_TOLERANCE_RPS)"
          },
          "half_width": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Half Width"
          },
          "rounds_needed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rounds Needed"
          },
          "rounds_missing": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rounds Missing"
          }
        },
        "type": "object",
        "required": [
          "rounds",
          "matches",
          "tolerance"
        ],
        "title": "Significance",
        "description": "How far the track record is from meaning something.\n\nThe bootstrap resamples ROUNDS, not matches: the ten matches of a round are\npredicted by the same fit, and treating them as independent would shrink\nthe interval by about sqrt(10) and make anything look significant.\nUnder four rounds there is nothing to resample, and the method says so."
      },
...
      "Standings": {
        "properties": {
          "season": {
            "type": "string",
            "title": "Season"
          },
          "matches_played": {
            "type": "integer",
            "title": "Matches Played"
          },
          "teams": {
            "type": "integer",
            "title": "Teams"
          },
          "last_match_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Match Date"
          },
          "tie_break": {
            "type": "string",
            "title": "Tie Break"
          },
          "table": {
            "items": {
              "$ref": "#/components/schemas/StandingRow"
            },
            "type": "array",
            "title": "Table",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "season",
          "matches_played",
          "teams",
          "tie_break"
        ],
        "title": "Standings",
        "description": "La classifica: aritmetica sui risultati, non una previsione.\n\n`tie_break` dichiara come sono separate le squadre a pari punti, perche' in\nSerie A contano prima gli scontri diretti e non la differenza reti — una\ntabella ordinata nel modo sbagliato sembra giusta proprio nelle settimane\nin cui qualcuno la guarda."
      },
...
      "TrackRecord": {
        "properties": {
          "season": {
            "type": "string",
            "title": "Season"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "rps_backtest_reference": {
            "type": "number",
            "title": "Rps Backtest Reference"
          },
          "by_round": {
            "items": {
              "$ref": "#/components/schemas/RoundPoint"
            },
            "type": "array",
            "title": "By Round",
            "default": []
          },
          "by_match": {
            "items": {
              "$ref": "#/components/schemas/MatchPoint"
            },
            "type": "array",
            "title": "By Match",
            "default": []
          },
          "calibration": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CalibrationBin"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibration",
            "description": "Null below 50 resolved predictions: fewer points make a curve that is noise drawn nicely"
          },
          "calibration_unavailable_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibration Unavailable Reason"
          }
        },
        "type": "object",
        "required": [
          "season",
          "rps_backtest_reference"
        ],
        "title": "TrackRecord"
      },
...
    }
  }
}